Work in progress
This started off as an assignment but it was kind of bland and I knew I could alter it to create a better user experience. 
Future commits will add the ability to set the number of questions per quiz, select one or more difficulty levels, select one or more topics, and the questions will be given in random order. Expected to be completed by 7-10-22